% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_transform_string.R
\name{clean_transform_string}
\alias{clean_transform_string}
\title{Clean and Transform String}
\usage{
clean_transform_string(x)
}
\arguments{
\item{x}{Character vector to be cleaned and transformed.}
}
\value{
Cleaned and transformed character vector.
}
\description{
This function cleans and transforms a given string by:
\enumerate{
\item Removing unnecessary whitespaces.
\item Removing non-ASCII characters.
\item Converting to lowercase.
\item Removing all non-alphanumeric characters.
}
}
\examples{
\dontrun{
clean_transform_string("  Hello Wörld!  ")
}
}
