% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rut_check.R
\name{compute_check_digit}
\alias{compute_check_digit}
\title{Compute the Check Digit for a Chilean RUT}
\usage{
compute_check_digit(rut)
}
\arguments{
\item{rut}{The RUT number as a numeric or character string. Dots and characters after '-' in the string will be removed.}
}
\value{
Returns the calculated check digit as a character string.
}
\description{
This function calculates the check digit for a given RUT (Rol Único Tributario),
which is a Chilean tax identification number.
}
\examples{
rut_number <- "76086428"
compute_check_digit(rut_number)

rut_number_with_dots <- "76.086.428"
compute_check_digit(rut_number_with_dots)

rut_with_hyphen <- "76.086.428-K"
compute_check_digit(rut_with_hyphen)

}
