% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_bigmark.R
\name{format_bigmark}
\alias{format_bigmark}
\title{Format Numbers with Big Marks}
\usage{
format_bigmark(x, digits = 3)
}
\arguments{
\item{x}{Numeric value to be formatted.}

\item{digits}{Number of digits to round to. Default is 3.}
}
\value{
Formatted string.
}
\description{
This function takes a numeric value and formats it using big marks (commas)
for better readability. It also allows for specifying the number of decimal places.
}
\examples{
format_bigmark(1234567.89)
format_bigmark(1234567.89, digits = 2)

}
