% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_table.R
\name{pdf_table}
\alias{pdf_table}
\title{pdf_table}
\usage{
pdf_table(table, file_name)
}
\arguments{
\item{table}{This is the table we want to save}

\item{file_name}{This is the name of the file that we want to create}
}
\description{
This function produces a .tex file for tables and print a pdf of the table (usually regression or summary tables)
}
