% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_summary.R
\name{model_summary}
\alias{model_summary}
\title{Model Summary in LaTeX Format}
\usage{
model_summary(reg, output, notes, coef_map)
}
\arguments{
\item{reg}{Regression object for which summary is generated.}

\item{output}{Output format, although the function currently only supports LaTeX.}

\item{notes}{Notes to include at the bottom of the table.}

\item{coef_map}{List of coefficients to omit from the summary.}
}
\value{
A LaTeX-formatted summary of the regression model.
}
\description{
This function creates a summary of a regression model and outputs it in LaTeX format.
The summary includes various statistics like \( R^2 \), adjusted \( R^2 \), and others.
}
\examples{
\dontrun{
# Assuming 'model' is a regression object
model_summary(model, "latex", list_names = c("Intercept" = "Constant"),
              notes = "Standard errors in parentheses.",
              rows = NULL, coef_omit_list = NULL)
}
}
